Update playsound.py
Changing some error stuff, and experimenting with the block argument in play sound function. (If block is True, it will block every process running, and will wait until the media has stopped playing, if not, it will just run asynchronously. But this is not working for me so I'm experimenting a little bit.